refactor(strm-1393): rename member package to user
