+++ pkg/entity/user/cmd.go
@@ -1,4 +1,4 @@
-package member
+package user
 
 import (
 	"github.com/spf13/cobra"
