+++ pkg/entity/user/printers.go
@@ -1,4 +1,4 @@
-package member
+package user
 
 import (
 	"errors"
