+++ pkg/entity/user/user.go
@@ -1,4 +1,4 @@
-package member
+package user
 
 import (
 	"context"
